# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CheapestOut": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CheapestPin"
            },
            "title": "Items",
            "type": "array"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "items",
          "unit"
        ],
        "title": "CheapestOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OfferList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Offer"
            },
            "title": "Items",
            "type": "array"
          },
          "stale_thresholds": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Stale Thresholds",
            "type": "object"
          }
        },
        "required": [
          "items",
          "stale_thresholds"
        ],
        "title": "OfferList",
        "type": "object"
      },
...
      "VoidIn": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "VoidIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kitchen ERP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ingredients/{ingredient_id}/offers": {
      "get": {
        "operationId": "ingredient_offers_api_v1_ingredients__ingredient_id__offers_get",
        "parameters": [
          {
            "in": "path",
            "name": "ingredient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ingredient Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "min_quality",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 5,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Quality"
            }
          },
          {
            "in": "query",
            "name": "exclude_stale",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exclude Stale",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "exclude_promo",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exclude Promo",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingredient Offers",
        "tags": [
          "purchases"
        ]
      }
    },
...
    "/api/v1/price-book/cheapest": {
      "get": {
        "operationId": "cheapest_api_v1_price_book_cheapest_get",
        "parameters": [
          {
            "in": "query",
            "name": "ingredient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ingredient Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "min_quality",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 5,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Quality"
            }
          },
          {
            "in": "query",
            "name": "exclude_stale",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exclude Stale",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheapestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cheapest",
        "tags": [
          "purchases"
        ]
      }
    },
    "/api/v1/price-book/compare": {
      "post": {
        "operationId": "compare_api_v1_price_book_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare",
        "tags": [
          "purchases"
        ]
      }
    },
...
    "/api/v1/price-observations": {
      "get": {
        "operationId": "list_observations_api_v1_price_observations_get",
        "parameters": [
          {
            "in": "query",
            "name": "product_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "in": "query",
            "name": "vendor_location_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vendor Location Id"
            }
          },
          {
            "in": "query",
            "name": "include_voided",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Voided",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Observations",
        "tags": [
          "purchases"
        ]
      },
      "post": {
        "description": "A shelf price: the simplest producer of observations (capture contract).",
        "operationId": "create_observation_api_v1_price_observations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Observation",
        "tags": [
          "purchases"
        ]
      }
    },
    "/api/v1/price-observations/{observation_id}": {
      "get": {
        "operationId": "get_observation_api_v1_price_observations__observation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Observation",
        "tags": [
          "purchases"
        ]
      }
    },
    "/api/v1/price-observations/{observation_id}/void": {
      "post": {
        "operationId": "void_observation_api_v1_price_observations__observation_id__void_post",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Void Observation",
        "tags": [
          "purchases"
        ]
      }
    },
...
    "/api/v1/products/{product_id}/last-purchase-unit": {
      "get": {
        "operationId": "last_purchase_unit_api_v1_products__product_id__last_purchase_unit_get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LastUnitOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Last Purchase Unit",
        "tags": [
          "purchases"
        ]
      }
    },
    "/api/v1/products/{product_id}/prices": {
      "get": {
        "operationId": "product_prices_api_v1_products__product_id__prices_get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Product Prices",
        "tags": [
          "purchases"
        ]
      }
    },
...
    "/api/v1/purchases/{purchase_id}/commit": {
      "post": {
        "operationId": "commit_purchase_api_v1_purchases__purchase_id__commit_post",
        "parameters": [
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Purchase Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit Purchase",
        "tags": [
          "purchases"
        ]
      }
    },
    "/api/v1/purchases/{purchase_id}/lines": {
      "post": {
        "operationId": "add_line_api_v1_purchases__purchase_id__lines_post",
        "parameters": [
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Purchase Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Line",
        "tags": [
          "purchases"
        ]
      }
    },
...
    "/api/v1/purchases/{purchase_id}/lines/{line_id}/re-resolve": {
      "post": {
        "operationId": "re_resolve_line_api_v1_purchases__purchase_id__lines__line_id__re_resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Purchase Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "line_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Line Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Re Resolve Line",
        "tags": [
          "purchases"
        ]
      }
    },
...
  }
}